requirements.txt
streamlit
pymupdf
torch==2.2.2
sentence-transformers==2.7.0
transformers==4.41.2
accelerate==0.30.1
numpy<2
weaviate-client
openai
python-dotenv
requests
streamlit-mermaid
dspy-ai
ragas
tenacity
pyyaml